Move  helper functions such as safe memory management and string management into the Util.c module
