+++ parser/src/SemanticChecker.c
@@ -1,95 +1,10 @@
 /* Adapted from Skeleton.c generated by the BNFC Converter  */
 /*** Semantic Error Checking using Visitor Traversal. ***/
 
-#include "SemanticChecker.h" // Include the header above
-
-
-// --- Helper Functions ---
-
-
-static void free_safe(void* ptr) {
-    if (ptr) free(ptr);
-}
-
-
-static void* malloc_safe(size_t size) {
-    void* ptr = malloc(size);
-    if (!ptr) {
-        perror("Failed to allocate memory");
-        exit(EXIT_FAILURE);
-    }
-    return ptr;
-}
-
-
-// Helper duplicate strings safely
-static char* strdup_safe(const char* src) {
-    if (!src) return NULL;
-    size_t len = strlen(src);
-    char* dst = malloc_safe(len + 1);
-    if (dst) strcpy(dst, src);
-    return dst;
-}
-
-
-// Helper to format strings with variable arguments
-char *format_string(const char *fmt, ...) {
-    va_list args;
-    va_start(args, fmt);
-
-    // Calculate required length
-    int needed = vsnprintf(NULL, 0, fmt, args);
-    va_end(args);
-
-    // Formatting error
-    if (needed < 0) {
-        va_end(args);
-        return NULL; 
-    }
-
-    // Allocate buffer
-    char *buffer = malloc_safe(needed + 1);
-
-    // Format string
-    va_start(args, fmt);
-    vsnprintf(buffer, needed + 1, fmt, args);
-    va_end(args);
-
-    return buffer;
-}
-
-
-// Helper to append formatted strings to a buffer
-char *appendf(char *buffer, size_t *size, size_t *used, const char *fmt, ...) {
-    va_list args;
-    va_start(args, fmt);
-
-    // Calculate required length
-    int needed = vsnprintf(NULL, 0, fmt, args);
-    va_end(args);
-
-    if (*used + needed + 1 >= *size) {
-        *size = (*used + needed + 1) * 2; // Grow size of buffer
-        buffer = realloc(buffer, *size);
-        if (!buffer) {
-            perror("Failed to reallocate memory for buffer");
-            exit(EXIT_FAILURE);
-        }
-    }
-
-    va_start(args, fmt);
-    vsnprintf(buffer + *used, *size - *used, fmt, args);
-    va_end(args);
-
-    *used += needed;
-    return buffer;
-}
+#include "SemanticChecker.h"
 
 
 // --- Error Reporting ---
-
-#define ERR_INITIAL_CAPACITY 4
-
 
 // Helper to intitialise the error list
 void initErrorList(SemanticContext *ctx) {
@@ -157,6 +72,7 @@ void reportError(SemanticContext *ctx, const char *format, ...) {
 }
 
 
+// Helper to convert error codes to strings
 const char* errorCodeToString(ErrorCode code) {
     switch (code) {
         case MultipleDeclaration: return "MultipleDeclaration";
@@ -170,6 +86,7 @@ const char* errorCodeToString(ErrorCode code) {
 }
 
 
+// Helper to return errors as a string in a human-readable format
 char *reportErrors(SemanticContext *ctx) {
     size_t size = 1024;
     size_t used = 0;
@@ -179,14 +96,14 @@ char *reportErrors(SemanticContext *ctx) {
     if (ctx && ctx->errorCount > 0) {
         for (int i = 0; i < ctx->errorCount; i++) {
             VNNLibError e = ctx->errors[i];
-            buffer = appendf(buffer, &size, &used, 
+            buffer = append_str(buffer, &size, &used, 
                 "[%s] %s (symbol: %s)\nHint: %s\n", 
                 errorCodeToString(e.errorCode),
                 e.message, 
                 e.offendingSymbol, 
                 e.hint);
 
-            buffer = appendf(buffer, &size, &used, "\n");
+            buffer = append_str(buffer, &size, &used, "\n");
         }
     }
 
@@ -194,18 +111,19 @@ char *reportErrors(SemanticContext *ctx) {
 }
 
 
+// Helper to return errors as a JSON string
 char *reportErrorsJSON(SemanticContext *ctx) {
     size_t size = 1024;
     size_t used = 0;
     char *buffer = malloc_safe(size);
     buffer[0] = '\0';
-    buffer = appendf(buffer, &size, &used, "{\n  \"errors\": [\n");
+    buffer = append_str(buffer, &size, &used, "{\n  \"errors\": [\n");
 
 
     if (ctx && ctx->errorCount > 0) {
         for (int i = 0; i < ctx->errorCount; i++) {
             VNNLibError e = ctx->errors[i];
-            buffer = appendf(buffer, &size, &used,
+            buffer = append_str(buffer, &size, &used,
                 "    {\n"
                 "      \"message\": \"%s\",\n"
                 "      \"offendingSymbol\": \"%s\",\n"
@@ -218,13 +136,13 @@ char *reportErrorsJSON(SemanticContext *ctx) {
                 errorCodeToString(e.errorCode));
 
             if (i < ctx->errorCount - 1) {
-                buffer = appendf(buffer, &size, &used, ",");
+                buffer = append_str(buffer, &size, &used, ",");
             }
-            buffer = appendf(buffer, &size, &used, "\n");
+            buffer = append_str(buffer, &size, &used, "\n");
         }
     }
 
-    buffer = appendf(buffer, &size, &used, "  ]\n}\n");
+    buffer = append_str(buffer, &size, &used, "  ]\n}\n");
     return buffer;
 }
 
@@ -328,32 +246,6 @@ SymbolInfo* findSymbol(SemanticContext *ctx, VariableName name) {
 }
 
 
-// Main entry point
-int checkSemantics(Query p) {
-    SemanticContext ctx;
-    initSemanticContext(&ctx);
-
-    checkQuery(p, &ctx); // Start traversal
-
-    // Check for semantic errors
-    if (ctx.errorCount > 0) {
-        char *errorReport = reportErrors(&ctx);
-        fprintf(stderr, "%s", errorReport);
-        free_safe(errorReport);
-    }
-
-    int finalErrorCount = ctx.errorCount;
-    destroySemanticContext(&ctx); // Clean up allocated symbols
-
-    if (finalErrorCount > 0) {
-        fprintf(stderr, "Found %d semantic error(s).\n", finalErrorCount);
-        return 1; // Indicate failure
-    } else {
-        return 0; // Indicate success
-    }
-}
-
-
 // ----------------- Recursive Traversal Functions -----------------
 
 // Check the ListInt production and add each discovered Int to the provided shape array
@@ -389,6 +281,7 @@ int checkArithExpr(ArithExpr p, SemanticContext *ctx)
     int err = 0;
     switch(p->kind)
     {
+        // TODO: Add type checking to ensure that the operands are of the same type
         case is_VarExpr:
             err |= checkTensorElement(p->u.varexpr_.tensorelement_, ctx);
             break;
@@ -405,16 +298,13 @@ int checkArithExpr(ArithExpr p, SemanticContext *ctx)
             err |= checkArithExpr(p->u.negate_.arithexpr_, ctx);
             break;
         case is_Plus:
-            // TODO Check for at least two elements in the list
             err |= checkListArithExpr(p->u.plus_.listarithexpr_, ctx);
             break;
         case is_Minus:
-			// TODO: Check for at least two elements in the list
             err |= checkArithExpr(p->u.minus_.arithexpr_, ctx);
             err |= checkListArithExpr(p->u.minus_.listarithexpr_, ctx);
             break;
         case is_Multiply:
-			// TODO: Check: Check for at least two elements in the list
             err |= checkListArithExpr(p->u.multiply_.listarithexpr_, ctx);
             break;
 
@@ -725,7 +615,6 @@ int checkQuery(Query p, SemanticContext *ctx)
 
 
 // --- Base Type Checkers ---
-// Return err if the base type is not NULL
 
 int checkSDouble(SDouble p, SemanticContext *ctx) { return 0; }
 int checkSInt(SInt p, SemanticContext *ctx) { return 0; }
@@ -741,7 +630,6 @@ int checkString(String s, SemanticContext *ctx) { return 0; }
 int checkScalar(SymbolInfo *symbol, TensorElement element, SemanticContext *ctx, char *tensorDim) {
     int index = strtol(tensorDim, NULL, 10);
     if (index != 0) {
-        // reportError(ctx, "Invalid tensor element '%s'. Expected index 0 for scalar variable.", element);
         addError(ctx, (VNNLibError) {
             .message = "Invalid tensor element",
             .offendingSymbol = element,
@@ -766,7 +654,6 @@ int checkTensor(SymbolInfo *symbol, TensorElement element, SemanticContext *ctx,
     while (idx < symbol->numDimensions) {
         if (token == NULL) {
             err = 1;
-            // reportError(ctx, "Not enough indices provided for tensor element '%s'.", element);
             addError(ctx, (VNNLibError) {
                 .message = "Not enough indices provided",
                 .offendingSymbol = element,
@@ -783,7 +670,6 @@ int checkTensor(SymbolInfo *symbol, TensorElement element, SemanticContext *ctx,
 
     if (token != NULL) {
         err = 1;
-        // reportError(ctx, "Too many indices provided for tensor element '%s'.", element);
         addError(ctx, (VNNLibError) {
             .message = "Too many indices provided",
             .offendingSymbol = element,
@@ -826,7 +712,6 @@ int checkTensorElement(TensorElement p, SemanticContext *ctx) {
 
     // Check that the tensor indices are provided
     if (!tensorDim) {
-        // reportError(ctx, "Invalid tensor element '%s'. Expected format: <tensor>_<index>", p);
         err = 1;
         addError(ctx, (VNNLibError) {
             .message = "Invalid tensor element",
@@ -841,7 +726,6 @@ int checkTensorElement(TensorElement p, SemanticContext *ctx) {
 
     // Check that variable is declared
     if (!symbol) {
-        // reportError(ctx, "Variable '%s' used but not declared.", p);
         err = 1;
         addError(ctx, (VNNLibError) {
             .message = "Undeclared variable",
